Got things into a close to working state with AK4 files, still work to be done

+++ AudioKit.playgroundbook/Contents/Sources/Internals/AKTable.swift
@@ -152,7 +152,7 @@ import Foundation
     public var phaseOffset: Int {
         @inline(__always)
         get {
-            return Int(phase * Double(count)
+            return Int(phase * Float(count))
         }
     }
